citibike-project/pyproject.toml
[project]
name = "citibike_project"
version = "0.0.1"
authors = [{ name = "kisarawenu@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "build>=1.3.0",
    "databricks-connect==16.4.10",
    # Any dependencies for jobs and pipelines in this project can be added here
    # See also https://docs.databricks.com/dev-tools/bundles/library-dependencies
    #
    # LIMITATION: for pipelines, dependencies are cached during development;
    # add dependencies to the 'environment' section of your pipeline.yml file instead
    "databricks-sdk>=0.74.0",
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
    "ruff>=0.14.9",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect==16.4.*",
]

[project.scripts]
main = "citibike_project.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/citibike_project"]

[tool.black]
line-length = 125
